refactor: Store- common 폴더 생성
