+++ src/components/common/Header.tsx
@@ -5,7 +5,7 @@ import { Back } from '@/assets/icons';
 import mq from '@/utils/mediaquery';
 import theme from '@/styles/theme';
 import Icons from '@/components/common/Icons';
-import usePagesStore from '@/stores/usePagesStore';
+import usePagesStore from '@/stores/common/usePagesStore';
 
 function Header({
   title,
