+++ src/pages/lunch/index.tsx
@@ -6,7 +6,7 @@ import LunchMainContents from '@/components/lunch/LunchMainContents';
 import LunchReviewsByRanking from '@/components/lunch/review/LunchReviewsByRanking';
 import COMPONENT_NAME from '@/constants/common/pages';
 import { HOME_TEXTS } from '@/constants/home/homeTexts';
-import usePagesStore from '@/stores/usePagesStore';
+import usePagesStore from '@/stores/common/usePagesStore';
 import theme from '@/styles/theme';
 import { css } from '@emotion/react';
